Refactor: consolidate test data and move mocks
