+++ test/pages/settings/advanced_settings_page_test.dart
@@ -7,8 +7,8 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
 
+import '../../mocks.dart';
 import '../../widget_test_utils.dart';
-import 'mocks.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
