+++ test/pages/settings/dashboards/dashboard_definition_page_test.dart
@@ -16,9 +16,9 @@ import 'package:lotti/widgets/sync/imap_config_utils.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:mocktail/mocktail.dart';
 
+import '../../../mocks.dart';
 import '../../../test_data.dart';
 import '../../../widget_test_utils.dart';
-import '../mocks.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
