+++ test/pages/tasks/tasks_page_test.dart
@@ -17,10 +17,9 @@ import 'package:lotti/themes/themes_service.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:mocktail/mocktail.dart';
 
-import '../../data/test_data.dart';
+import '../../mocks.dart';
 import '../../test_data.dart';
 import '../../widget_test_utils.dart';
-import '../settings/mocks.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
